[all] Making Network compilable, renamed USB device in 4S/4E/G55
