--- cores/arduino/arch_common/Server.hpp
@@ -1,30 +1,0 @@
-/*
-  Server.h - Base class that provides Server
-  Copyright (c) 2011 Adrian McEwen.  All right reserved.
-
-  This library is free software; you can redistribute it and/or
-  modify it under the terms of the GNU Lesser General Public
-  License as published by the Free Software Foundation; either
-  version 2.1 of the License, or (at your option) any later version.
-
-  This library is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-  Lesser General Public License for more details.
-
-  You should have received a copy of the GNU Lesser General Public
-  License along with this library; if not, write to the Free Software
-  Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
-*/
-
-#ifndef _ARDUINO_CORE_NETWORK_SERVER_HPP_
-#define _ARDUINO_CORE_NETWORK_SERVER_HPP_
-
-#include "CorePrint.hpp"
-
-class Server : public Print {
-public:
-  virtual void begin() =0;
-};
-
-#endif // _ARDUINO_CORE_NETWORK_SERVER_HPP_
