+++ tools/CMSIS_Devices/ATMEL/sam4s/include/sam4s8b.h
@@ -41,7 +41,7 @@
 
 #ifdef __cplusplus
  extern "C" {
-#endif 
+#endif
 
 #include <stdint.h>
 
@@ -91,7 +91,7 @@ typedef enum IRQn
   PWM_IRQn             = 31, /**< 31 SAM4S8B Pulse Width Modulation (PWM) */
   CRCCU_IRQn           = 32, /**< 32 SAM4S8B CRC Calculation Unit (CRCCU) */
   ACC_IRQn             = 33, /**< 33 SAM4S8B Analog Comparator (ACC) */
-  UDP_IRQn             = 34, /**< 34 SAM4S8B USB Device Port (UDP) */
+  USBDEV_IRQn          = 34, /**< 34 SAM4S8B USB Device Port (USBDEV) */
 
   PERIPH_COUNT_IRQn    = 35  /**< Number of peripheral IDs */
 } IRQn_Type;
@@ -100,7 +100,7 @@ typedef struct _DeviceVectors
 {
   /* Stack pointer */
   void* pvStack;
-  
+
   /* Cortex-M handlers */
   void* pfnReset_Handler;
   void* pfnNMI_Handler;
@@ -153,7 +153,7 @@ typedef struct _DeviceVectors
   void* pfnPWM_Handler;    /* 31 Pulse Width Modulation */
   void* pfnCRCCU_Handler;  /* 32 CRC Calculation Unit */
   void* pfnACC_Handler;    /* 33 Analog Comparator */
-  void* pfnUDP_Handler;    /* 34 USB Device Port */
+  void* pfnUSBDEV_Handler; /* 34 USB Device Port */
 } DeviceVectors;
 
 /* Cortex-M4 core handlers */
@@ -192,13 +192,13 @@ void TWI0_Handler       ( void );
 void TWI1_Handler       ( void );
 void UART0_Handler      ( void );
 void UART1_Handler      ( void );
-void UDP_Handler        ( void );
+void USBDEV_Handler     ( void );
 void USART0_Handler     ( void );
 void USART1_Handler     ( void );
 void WDT_Handler        ( void );
 
 /**
- * \brief Configuration of the Cortex-M4 Processor and Core Peripherals 
+ * \brief Configuration of the Cortex-M4 Processor and Core Peripherals
  */
 
 #define __CM4_REV              0x0001 /**< SAM4S8B core revision number ([15:8] revision number, [7:0] patch number) */
@@ -246,7 +246,7 @@ void WDT_Handler        ( void );
 #include "component/tc.h"
 #include "component/twi.h"
 #include "component/uart.h"
-#include "component/udp.h"
+#include "component/usbdev.h"
 #include "component/usart.h"
 #include "component/wdt.h"
 /*@}*/
@@ -283,7 +283,7 @@ void WDT_Handler        ( void );
 #define ID_PWM    (31) /**< \brief Pulse Width Modulation (PWM) */
 #define ID_CRCCU  (32) /**< \brief CRC Calculation Unit (CRCCU) */
 #define ID_ACC    (33) /**< \brief Analog Comparator (ACC) */
-#define ID_UDP    (34) /**< \brief USB Device Port (UDP) */
+#define ID_USBDEV (34) /**< \brief USB Device Port (USBDEV) */
 
 #define ID_PERIPH_COUNT (35) /**< \brief Number of peripheral IDs */
 /*@}*/
@@ -311,7 +311,7 @@ define HSMCI      ((Hsmci  *)0x40000000U) /**< \brief (HSMCI     ) Base Address 
 #define PDC_USART0 ((Pdc    *)0x40024100U) /**< \brief (PDC_USART0) Base Address */
 #define USART1     ((Usart  *)0x40028000U) /**< \brief (USART1    ) Base Address */
 #define PDC_USART1 ((Pdc    *)0x40028100U) /**< \brief (PDC_USART1) Base Address */
-#define UDP        ((Udp    *)0x40034000U) /**< \brief (UDP       ) Base Address */
+#define USBDEV     ((USBDev *)0x40034000U) /**< \brief (USBDEV    ) Base Address */
 #define ADC        ((Adc    *)0x40038000U) /**< \brief (ADC       ) Base Address */
 #define PDC_ADC    ((Pdc    *)0x40038100U) /**< \brief (PDC_ADC   ) Base Address */
 #define DACC       ((Dacc   *)0x4003C000U) /**< \brief (DACC      ) Base Address */
@@ -417,7 +417,7 @@ define HSMCI      ((Hsmci  *)0x40000000U) /**< \brief (HSMCI     ) Base Address 
 #endif
 
 /* HYSTeresis levels: please refer to Electrical Characteristics */
-#define ACC_ACR_HYST_50MV_MAX	          (0x01UL)
+#define ACC_ACR_HYST_50MV_MAX	        (0x01UL)
 #define ACC_ACR_HYST_90MV_MAX           (0x11UL)
 
 #ifdef __cplusplus
