+++ tools/CMSIS_Devices/ATMEL/sam4s/include/sam4sd32c.h
@@ -41,7 +41,7 @@
 
 #ifdef __cplusplus
  extern "C" {
-#endif 
+#endif
 
 #include <stdint.h>
 
@@ -64,7 +64,7 @@ typedef enum IRQn
   PendSV_IRQn           = -2,  /**< 14 Cortex-M4 Pend SV Interrupt           */
   SysTick_IRQn          = -1,  /**< 15 Cortex-M4 System Tick Interrupt       */
 /******  SAM4SD32C specific Interrupt Numbers *********************************/
-  
+
   SUPC_IRQn            =  0, /**<  0 SAM4SD32C Supply Controller (SUPC) */
   RSTC_IRQn            =  1, /**<  1 SAM4SD32C Reset Controller (RSTC) */
   RTC_IRQn             =  2, /**<  2 SAM4SD32C Real Time Clock (RTC) */
@@ -96,7 +96,7 @@ typedef enum IRQn
   PWM_IRQn             = 31, /**< 31 SAM4SD32C Pulse Width Modulation (PWM) */
   CRCCU_IRQn           = 32, /**< 32 SAM4SD32C CRC Calculation Unit (CRCCU) */
   ACC_IRQn             = 33, /**< 33 SAM4SD32C Analog Comparator (ACC) */
-  UDP_IRQn             = 34, /**< 34 SAM4SD32C USB Device Port (UDP) */
+  USBDEV_IRQn          = 34, /**< 34 SAM4SD32C USB Device Port (USBDEV) */
 
   PERIPH_COUNT_IRQn    = 35  /**< Number of peripheral IDs */
 } IRQn_Type;
@@ -105,7 +105,7 @@ typedef struct _DeviceVectors
 {
   /* Stack pointer */
   void* pvStack;
-  
+
   /* Cortex-M handlers */
   void* pfnReset_Handler;
   void* pfnNMI_Handler;
@@ -158,7 +158,7 @@ typedef struct _DeviceVectors
   void* pfnPWM_Handler;    /* 31 Pulse Width Modulation */
   void* pfnCRCCU_Handler;  /* 32 CRC Calculation Unit */
   void* pfnACC_Handler;    /* 33 Analog Comparator */
-  void* pfnUDP_Handler;    /* 34 USB Device Port */
+  void* pfnUSBDEV_Handler; /* 34 USB Device Port */
 } DeviceVectors;
 
 /* Cortex-M4 core handlers */
@@ -202,13 +202,13 @@ void TWI0_Handler       ( void );
 void TWI1_Handler       ( void );
 void UART0_Handler      ( void );
 void UART1_Handler      ( void );
-void UDP_Handler        ( void );
+void USBDEV_Handler     ( void );
 void USART0_Handler     ( void );
 void USART1_Handler     ( void );
 void WDT_Handler        ( void );
 
 /**
- * \brief Configuration of the Cortex-M4 Processor and Core Peripherals 
+ * \brief Configuration of the Cortex-M4 Processor and Core Peripherals
  */
 
 #define __CM4_REV              0x0001 /**< SAM4SD32C core revision number ([15:8] revision number, [7:0] patch number) */
@@ -258,7 +258,7 @@ void WDT_Handler        ( void );
 #include "component/tc.h"
 #include "component/twi.h"
 #include "component/uart.h"
-#include "component/udp.h"
+#include "component/usbdev.h"
 #include "component/usart.h"
 #include "component/wdt.h"
 /*@}*/
@@ -301,7 +301,7 @@ void WDT_Handler        ( void );
 #define ID_PWM    (31) /**< \brief Pulse Width Modulation (PWM) */
 #define ID_CRCCU  (32) /**< \brief CRC Calculation Unit (CRCCU) */
 #define ID_ACC    (33) /**< \brief Analog Comparator (ACC) */
-#define ID_UDP    (34) /**< \brief USB Device Port (UDP) */
+#define ID_USBDEV (34) /**< \brief USB Device Port (USBDEV) */
 
 #define ID_PERIPH_COUNT (35) /**< \brief Number of peripheral IDs */
 /*@}*/
@@ -330,7 +330,7 @@ void WDT_Handler        ( void );
 #define PDC_USART0 ((Pdc    *)0x40024100U) /**< \brief (PDC_USART0) Base Address */
 #define USART1     ((Usart  *)0x40028000U) /**< \brief (USART1    ) Base Address */
 #define PDC_USART1 ((Pdc    *)0x40028100U) /**< \brief (PDC_USART1) Base Address */
-#define UDP        ((Udp    *)0x40034000U) /**< \brief (UDP       ) Base Address */
+#define USBDEV     ((USBDev *)0x40034000U) /**< \brief (USBDEV    ) Base Address */
 #define ADC        ((Adc    *)0x40038000U) /**< \brief (ADC       ) Base Address */
 #define PDC_ADC    ((Pdc    *)0x40038100U) /**< \brief (PDC_ADC   ) Base Address */
 #define DACC       ((Dacc   *)0x4003C000U) /**< \brief (DACC      ) Base Address */
@@ -403,6 +403,7 @@ void WDT_Handler        ( void );
 
 #define CHIP_JTAGID       (0x05B3203FUL)
 #define CHIP_CIDR         (0x29A70EE0UL)
+#define CHIP_EXID         (0x0UL)
 #define NB_CH_ADC         (15UL)
 #define NB_CH_DAC         (2UL)
 #define USB_DEVICE_MAX_EP (8UL)
@@ -447,7 +448,7 @@ void WDT_Handler        ( void );
 #endif
 
 /* HYSTeresis levels: please refer to Electrical Characteristics */
-#define ACC_ACR_HYST_50MV_MAX	          (0x01UL)
+#define ACC_ACR_HYST_50MV_MAX	        (0x01UL)
 #define ACC_ACR_HYST_90MV_MAX           (0x11UL)
 
 #ifdef __cplusplus
