+++ tools/CMSIS_Devices/ATMEL/samg55/include/samg55j19.h
@@ -43,9 +43,7 @@
  extern "C" {
 #endif
 
-#if !(defined(__ASSEMBLY__) || defined(__IAR_SYSTEMS_ASM__))
 #include <stdint.h>
-#endif
 
 /* ************************************************************************** */
 /*   CMSIS DEFINITIONS FOR SAMG55J19 */
@@ -74,13 +72,13 @@ typedef enum IRQn
   WDT_IRQn             =  4, /**<  4 SAMG55J19 Watchdog Timer (WDT) */
   PMC_IRQn             =  5, /**<  5 SAMG55J19 Power Management Controller (PMC) */
   EFC0_IRQn            =  6, /**<  6 SAMG55J19 Enhanced Flash Controller (EFC0) */
-  FLEXCOM7_IRQn        =  7, /**< 7 SAMG55J19 FLEXCOM 7 (FLEXCOM7) */
+  FLEXCOM7_IRQn        =  7, /**<  7 SAMG55J19 FLEXCOM 7 (FLEXCOM7) */
   FLEXCOM0_IRQn        =  8, /**<  8 SAMG55J19 FLEXCOM 0 (FLEXCOM0) */
   FLEXCOM1_IRQn        =  9, /**<  9 SAMG55J19 FLEXCOM 1 (FLEXCOM1) */
   PIOA_IRQn            = 11, /**< 11 SAMG55J19 Parallel I/O Controller A (PIOA) */
   PIOB_IRQn            = 12, /**< 12 SAMG55J19 Parallel I/O Controller B (PIOB) */
   PDMIC0_IRQn          = 13, /**< 13 SAMG55J19 PDM 0 (PDMIC0) */
-  FLEXCOM2_IRQn        = 14,  /**< 14 SAMG55J19 FLEXCOM2 (FLEXCOM2) */
+  FLEXCOM2_IRQn        = 14, /**< 14 SAMG55J19 FLEXCOM2 (FLEXCOM2) */
   MEM2MEM_IRQn         = 15, /**< 15 SAMG55J19 MEM2MEM (MEM2MEM) */
   I2SC0_IRQn           = 16, /**< 16 SAMG55J19 I2SC0 (I2SC0) */
   I2SC1_IRQn           = 17, /**< 17 SAMG55J19 I2SC1 (I2SC1) */
@@ -96,7 +94,7 @@ typedef enum IRQn
   TC4_IRQn             = 27, /**< 27 SAMG55J19 Timer/Counter 4 (TC4) */
   TC5_IRQn             = 28, /**< 28 SAMG55J19 Timer/Counter 5 (TC5) */
   ADC_IRQn             = 29, /**< 29 SAMG55J19 Analog To Digital Converter (ADC) */
-  ARM_IRQn             = 30, /**< 30 SAMG55J19 FPU (ARM) */
+  FPU_IRQn             = 30, /**< 30 SAMG55J19 FPU (ARM) */
   WKUP0_IRQn           = 31, /**< 31 SAMG55J19 External interrupt 0 (WKUP0) */
   WKUP1_IRQn           = 32, /**< 32 SAMG55J19 External interrupt 1 (WKUP1) */
   WKUP2_IRQn           = 33, /**< 33 SAMG55J19 External interrupt 2 (WKUP2) */
@@ -114,7 +112,7 @@ typedef enum IRQn
   WKUP14_IRQn          = 45, /**< 45 SAMG55J19 External interrupt 14 (WKUP14) */
   WKUP15_IRQn          = 46, /**< 46 SAMG55J19 External interrupt 15 (WKUP15) */
   UHP_IRQn             = 47, /**< 47 SAMG55J19 USB OHCI (UHP) */
-  UDP_IRQn             = 48, /**< 48 SAMG55J19 USB Device FS (UDP) */
+  USBDEV_IRQn          = 48, /**< 48 SAMG55J19 USB Device FS (USBDEV) */
   CRCCU_IRQn           = 49, /**< 49 SAMG55J19 CRCCU */
   PERIPH_COUNT_IRQn    = 50  /**< Number of peripheral IDs */
 } IRQn_Type;
@@ -149,22 +147,22 @@ typedef struct _DeviceVectors
   void* pfnWDT_Handler;       /*  4 Watchdog Timer */
   void* pfnPMC_Handler;       /*  5 Power Management Controller */
   void* pfnEFC0_Handler;      /*  6 Enhanced Flash Controller */
-  void* pfnFLEXCOM7_Handler;  /* FLEXCOM7 */
-  void* pfnFLEXCOM0_Handler;  /* 8  FLEXCOM 0 */
-  void* pfnFLEXCOM1_Handler;  /* 9  FLEXCOM 1 */
+  void* pfnFLEXCOM7_Handler;  /*  7 FLEXCOM 7 */
+  void* pfnFLEXCOM0_Handler;  /*  8 FLEXCOM 0 */
+  void* pfnFLEXCOM1_Handler;  /*  9 FLEXCOM 1 */
   void* pvReserved10;
   void* pfnPIOA_Handler;      /* 11 Parallel I/O Controller A */
   void* pfnPIOB_Handler;      /* 12 Parallel I/O Controller B */
   void* pfnPDMIC0_Handler;    /* 13 PDM 0 */
-  void* pfnFLEXCOM2_Handler;  /* 14  FLEXCOM 2 */
+  void* pfnFLEXCOM2_Handler;  /* 14 FLEXCOM 2 */
   void* pfnMEM2MEM_Handler;   /* 15 MEM2MEM */
   void* pfnI2SC0_Handler;     /* 16 I2SC0 */
   void* pfnI2SC1_Handler;     /* 17 I2SC1 */
   void* pfnPDMIC1_Handler;    /* 18 PDM 1 */
-  void* pfnFLEXCOM3_Handler;  /* 19  FLEXCOM 3 */
-  void* pfnFLEXCOM4_Handler;  /* 20  FLEXCOM 4 */
-  void* pfnFLEXCOM5_Handler;  /* 21  FLEXCOM 5 */
-  void* pfnFLEXCOM6_Handler;  /* 22  FLEXCOM 6 */
+  void* pfnFLEXCOM3_Handler;  /* 19 FLEXCOM 3 */
+  void* pfnFLEXCOM4_Handler;  /* 20 FLEXCOM 4 */
+  void* pfnFLEXCOM5_Handler;  /* 21 FLEXCOM 5 */
+  void* pfnFLEXCOM6_Handler;  /* 22 FLEXCOM 6 */
   void* pfnTC0_Handler;       /* 23 Timer/Counter 0 */
   void* pfnTC1_Handler;       /* 24 Timer/Counter 1 */
   void* pfnTC2_Handler;       /* 25 Timer/Counter 2 */
@@ -172,7 +170,7 @@ typedef struct _DeviceVectors
   void* pfnTC4_Handler;       /* 27 Timer/Counter 4 */
   void* pfnTC5_Handler;       /* 28 Timer/Counter 5 */
   void* pfnADC_Handler;       /* 29 Analog To Digital Converter */
-  void* pfnARM_Handler;       /* 30 FPU */
+  void* pfnFPU_Handler;       /* 30 FPU */
   void* pvReserved31;
   void* pvReserved32;
   void* pvReserved33;
@@ -190,24 +188,8 @@ typedef struct _DeviceVectors
   void* pvReserved45;
   void* pvReserved46;
   void* pfnUHP_Handler;       /* 47 USB OHCI */
-  void* pfnUDP_Handler;       /* 48 USB Device FS */
-  void* pfnCRCCU_Handler;     /*49 CRCCU Device  */
-  void* pvReserved50;
-  void* pvReserved51;
-  void* pvReserved52;
-  void* pvReserved53;
-  void* pvReserved54;
-  void* pvReserved55;
-  void* pvReserved56;
-  void* pvReserved57;
-  void* pvReserved58;
-  void* pvReserved59;
-  void* pvReserved60;
-  void* pvReserved61;
-  void* pvReserved62;
-  void* pvReserved63;         /* UHP */
-  void* pvReserved64;         /* UDP */
-  void* pvReserved65;         /* CRCCU */
+  void* pfnUSBDEV_Handler;    /* 48 USB Device FS */
+  void* pfnCRCCU_Handler;     /* 49 CRCCU Device  */
 } DeviceVectors;
 
 /* Cortex-M4 core handlers */
@@ -245,16 +227,16 @@ void TC2_Handler        ( void );
 void TC3_Handler        ( void );
 void TC4_Handler        ( void );
 void TC5_Handler        ( void );
-void UDP_Handler        ( void );
+void USBDEV_Handler     ( void );
 void UHP_Handler        ( void );
-void FLEXCOM0_Handler      ( void );
-void FLEXCOM1_Handler      ( void );
-void FLEXCOM2_Handler      ( void );
-void FLEXCOM3_Handler      ( void );
-void FLEXCOM4_Handler      ( void );
-void FLEXCOM5_Handler      ( void );
-void FLEXCOM6_Handler      ( void );
-void FLEXCOM7_Handler      ( void );
+void FLEXCOM0_Handler   ( void );
+void FLEXCOM1_Handler   ( void );
+void FLEXCOM2_Handler   ( void );
+void FLEXCOM3_Handler   ( void );
+void FLEXCOM4_Handler   ( void );
+void FLEXCOM5_Handler   ( void );
+void FLEXCOM6_Handler   ( void );
+void FLEXCOM7_Handler   ( void );
 void WDT_Handler        ( void );
 /**
  * \brief Configuration of the Cortex-M4 Processor and Core Peripherals
@@ -304,7 +286,7 @@ void WDT_Handler        ( void );
 #include "component/supc.h"
 #include "component/tc.h"
 #include "component/twi.h"
-#include "component/udp.h"
+#include "component/usbdev.h"
 #include "component/uhp.h"
 #include "component/usart.h"
 #include "component/wdt.h"
@@ -347,7 +329,7 @@ void WDT_Handler        ( void );
 #define ID_ADC       (29) /**< \brief Analog To Digital Converter (ADC) */
 #define ID_ARM       (30) /**< \brief FPU (ARM) */
 #define ID_UHP       (47) /**< \brief USB OHCI (UHP) */
-#define ID_UDP       (48) /**< \brief USB Device FS (UDP) */
+#define ID_USBDEV    (48) /**< \brief USB Device FS (USBDEV) */
 #define ID_CRCCU     (49) /**< \brief CRCCU (CRCCU) */
 
 #define ID_PERIPH_COUNT (50) /**< \brief Number of peripheral IDs */
@@ -431,7 +413,7 @@ void WDT_Handler        ( void );
 #define PDC_SPI6    ((Pdc      *)0x40040500U) /**< \brief (PDC_SPI6   ) Base Address */
 #define TWI6        ((Twi      *)0x40040600U) /**< \brief (TWI6       ) Base Address */
 #define PDC_TWI6    ((Pdc      *)0x40040700U) /**< \brief (PDC_TWI6   ) Base Address */
-#define UDP         ((Udp      *)0x40044000U) /**< \brief (UDP        ) Base Address */
+#define USBDEV      ((USBDev   *)0x40044000U) /**< \brief (USBDEV     ) Base Address */
 #define CRCCU       ((Crccu    *)0x40048000U) /**< \brief (CRCCU      ) Base Address */
 #define UHP         ((Uhp      *)0x20400000U) /**< \brief (UHP        ) Base Address */
 #define MATRIX      ((Matrix   *)0x400E0200U) /**< \brief (MATRIX     ) Base Address */
@@ -494,7 +476,6 @@ void WDT_Handler        ( void );
 #define CHIP_FREQ_MAINCK_RC_24MHZ       (24000000UL)
 #define CHIP_FREQ_CPU_MAX               (120000000UL)
 #define CHIP_FREQ_XTAL_32K              (32768UL)
-
 
 /* Embedded Flash Write Wait State */
 #define CHIP_FLASH_WRITE_WAIT_STATE     (8U)
