+++ variants/atmel_sam4e_xplained_pro/variant.project
@@ -35,7 +35,7 @@
         <File Name="../../cores/arduino/arch_arm/core_main.cpp"/>
         <File Name="../../cores/arduino/arch_arm/core_shift.c"/>
         <File Name="../../cores/arduino/arch_arm/core_shift.h"/>
-        <File Name="../../cores/arduino/arch_arm/IPAddress.cpp"/>
+        <File Name="../../cores/arduino/arch_arm/CoreIPAddress.cpp"/>
         <File Name="../../cores/arduino/arch_arm/core_cortex_vectors.c"/>
         <File Name="../../cores/arduino/arch_arm/core_cortex_vectors.h"/>
         <File Name="../../cores/arduino/arch_arm/core_hooks.h"/>
@@ -44,7 +44,7 @@
         <File Name="../../cores/arduino/arch_arm/CorePrintable.hpp"/>
         <File Name="../../cores/arduino/arch_arm/CoreStream.hpp"/>
         <File Name="../../cores/arduino/arch_arm/CoreString.hpp"/>
-        <File Name="../../cores/arduino/arch_arm/IPAddress.hpp"/>
+        <File Name="../../cores/arduino/arch_arm/CoreIPAddress.hpp"/>
         <File Name="../../cores/arduino/arch_arm/core_math.cpp"/>
         <File Name="../../cores/arduino/arch_arm/core_math.hpp"/>
       </VirtualDirectory>
@@ -52,10 +52,10 @@
         <File Name="../../cores/arduino/arch_common/core_binary.h"/>
         <File Name="../../cores/arduino/arch_common/core_constants.h"/>
         <File Name="../../cores/arduino/arch_common/core_new.cpp"/>
-        <File Name="../../cores/arduino/arch_common/Client.hpp"/>
-        <File Name="../../cores/arduino/arch_common/Server.hpp"/>
+        <File Name="../../cores/arduino/arch_common/CoreClient.hpp"/>
+        <File Name="../../cores/arduino/arch_common/CoreServer.hpp"/>
         <File Name="../../cores/arduino/arch_common/core_new.hpp"/>
-        <File Name="../../cores/arduino/arch_common/Udp.hpp"/>
+        <File Name="../../cores/arduino/arch_common/CoreUdp.hpp"/>
         <File Name="../../cores/arduino/arch_common/core_reset.h"/>
         <File Name="../../cores/arduino/arch_common/core_watchdog.h"/>
       </VirtualDirectory>
@@ -184,36 +184,36 @@
     </Configuration>
   </Settings>
   <VirtualDirectory Name="product headers">
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/sam4e16e.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/system_sam4e.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/acc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/adc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/chipid.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/cmcc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/crccu.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/dacc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/efc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/gpbr.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/hsmci.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/matrix.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/pdc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/pio.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/pmc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/pwm.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/rstc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/rtc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/rtt.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/smc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/spi.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/ssc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/supc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/tc.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/twi.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/uart.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/udp.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/usart.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/component/wdt.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/sam4e.h"/>
-    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4s/include/pio/sam4e16e.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/sam4e16e.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/system_sam4e.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/acc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/adc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/chipid.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/cmcc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/crccu.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/dacc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/efc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/gpbr.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/hsmci.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/matrix.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/pdc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/pio.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/pmc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/pwm.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/rstc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/rtc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/rtt.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/smc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/spi.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/ssc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/supc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/tc.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/twi.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/uart.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/usbdev.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/usart.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/component/wdt.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/sam4e.h"/>
+    <File Name="../../tools/CMSIS_Devices/ATMEL/sam4e/include/pio/sam4e16e.h"/>
   </VirtualDirectory>
 </CodeLite_Project>
