+++ variants/atmel_sam4s_xplained/variant_startup.c
@@ -81,7 +81,7 @@ void DACC_Handler   ( void ) __attribute__ ((weak, alias("vector_halt")));
 void PWM_Handler    ( void ) __attribute__ ((weak, alias("vector_halt")));
 void CRCCU_Handler  ( void ) __attribute__ ((weak, alias("vector_halt")));
 void ACC_Handler    ( void ) __attribute__ ((weak, alias("vector_halt")));
-void UDP_Handler    ( void ) __attribute__ ((weak, alias("vector_halt")));
+void USBDEV_Handler ( void ) __attribute__ ((weak, alias("vector_halt")));
 
 /* Exception Table */
 __attribute__ ((section(".ram_isr_vector")))
@@ -167,7 +167,7 @@ DeviceVectors exception_table=
   .pfnPWM_Handler    = (void*) PWM_Handler,    /* 31 Pulse Width Modulation */
   .pfnCRCCU_Handler  = (void*) CRCCU_Handler,  /* 32 CRC Calculation Unit */
   .pfnACC_Handler    = (void*) ACC_Handler,    /* 33 Analog Comparator */
-  .pfnUDP_Handler    = (void*) UDP_Handler     /* 34 USB Device Port */
+  .pfnUSBDEV_Handler = (void*) USBDEV_Handler  /* 34 USB Device Port */
 };
 
 /**
